+++ Changelog.md
@@ -2,6 +2,10 @@
 
 _Also check out the [roadmap](Roadmap.md) with features I’d like to add in the future._
 
+### 29 May 2021 v1.2
+
+-   Support for Mega Chairman, if a host has both titles at once
+
 ### 28 May 2021 v1.1
 
 -   Support for Rickies statuses, in preparation for the [WWDC21 Keynote Rickies](https://rickies.co/keynote-jun-2021)